Decodify prep (#53)
* Clean up of xml DCCODE-112

* more clean up of xml DCCODE-112

* added annos on sections for organic split

added enacted="true" flag to all enacted titles

added enacted="false" to all sections that are too dificult
to move to their organic sections, and that we will therefore
leave in their current location in the code. the enacted="false"
will let us know that even though they are in the code, they
are not actually enacted.

added manual="true" to sections whose organic annotation is
just too "weird" to deal with with the parser. We will go
through and manually handle these organics.

* fixed section numbering typo

* more unenacted/manual sections

* fixed dates; added some law docs

* add instructions/docs for manual hist annos;

* added enacted="false" for all secs with no hist annos

* added <reason> tags; fixed cite comma

* removed all manual/effective; formatting

* fix effective dates

* manual hist fixes

* fix law

* switch back to container for title 2

* more manual fixes

* removed enacted="false" flags

* more fixes

* no validation errors