Fix insert instr. and placeholder-values:
The insert instruction at 7002(a) should automatically create a
placeholder with the language Applicable as of July 1, 2021.

The insert instruction at 7002(b) is a normal insert except, it needs
`placeholder-value=""` to not try to create a placeholder at para (1).
To create a placeholder showing a new (1A) you need the codify:placeholder
instruction.